feat:优化不必要的内链代码，添加适配Godot的代码
