+++ addons/GameplayTag/GameplayTagContainer.cs
@@ -1,15 +1,16 @@
-using System;
-using System.Collections.Generic;
 using Godot;
+using Godot.Collections;
 
 
 namespace Gameplay.Tag;
 
-
-public class GameplayTagContainer
+[GlobalClass]
+public partial class GameplayTagContainer : Resource
 {
-    public List<GameplayTag> GameplayTags { get; set; } = [];
-    public List<GameplayTag> ParentTags { get; set; } = [];
+    [Export]
+    public Array<GameplayTag> GameplayTags { get; set; } = [];
+    [Export]
+    public Array<GameplayTag> ParentTags { get; set; } = [];
 
 
     public bool IsEmpty()
@@ -26,7 +27,7 @@ public class GameplayTagContainer
     public GameplayTagContainer GetGameplayTagParents()
     {
         var result = new GameplayTagContainer();
-        result.GameplayTags = new List<GameplayTag>(GameplayTags);
+        result.GameplayTags = new Array<GameplayTag>(GameplayTags);
         for (var i = 0; i < ParentTags.Count; i++)
         {
             var tag = ParentTags[i];
@@ -250,103 +251,6 @@ public class GameplayTagContainer
     }
 }
 
-/// <summary>
-/// 生成后，就不会修改，所以同名的标签，可以一直使用一个GameplayTag对象
-/// </summary>
-public struct GameplayTag : IEquatable<GameplayTag>, IComparable<GameplayTag>
-{
-    public string TagName { get; private set; }
-
-    public GameplayTag(string tagName)
-    {
-        TagName = string.Intern(tagName);
-    }
-
-    public bool IsValid()
-    {
-        return !string.IsNullOrEmpty(TagName);
-    }
-
-
-    public bool MatchesTag(GameplayTag tagToCheck)
-    {
-        var completeContainer = GameplayTagsManager.Instance.GetSingleTagContainer(this);
-        if (completeContainer != null)
-        {
-            return completeContainer.HasTag(tagToCheck);
-        }
-        return false;
-    }
-
-    public bool MatchesTagExact(GameplayTag tagToCheck)
-    {
-        if (tagToCheck.IsValid())
-        {
-            return TagName == tagToCheck.TagName;
-        }
-
-        return false;
-    }
-
-    public bool MatchesAny(GameplayTagContainer containerToCheck)
-    {
-        var completeContainer = GameplayTagsManager.Instance.GetSingleTagContainer(this);
-        if (completeContainer != null)
-        {
-            return completeContainer.HasAny(containerToCheck);
-        }
-        return false;
-    }
-
-    public bool MatchesAnyExact(GameplayTagContainer containerToCheck)
-    {
-        if (!containerToCheck.IsEmpty())
-        {
-            return containerToCheck.GameplayTags.BinarySearch(this) >= 0;
-        }
-        return false;
-    }
-
-    public bool Equals(GameplayTag other)
-    {
-        return TagName == other.TagName;
-    }
-
-    public override bool Equals(object obj)
-    {
-        if (obj is GameplayTag other)
-        {
-            return TagName == other.TagName;
-        }
-        return false;
-    }
-
-    public override int GetHashCode()
-    {
-        return TagName?.GetHashCode() ?? 0;
-    }
-    
-    public static bool operator ==(GameplayTag left, GameplayTag right)
-    {
-        return left.Equals(right);
-    }
-
-    public static bool operator !=(GameplayTag left, GameplayTag right)
-    {
-        return !(left == right);
-    }
-    
-    public int CompareTo(GameplayTag other)
-    {
-        return string.Compare(TagName, other.TagName, StringComparison.OrdinalIgnoreCase);
-    }
-    
-    public override string ToString()
-    {
-        return TagName;
-    }
-}
-
 
 public enum GameplayTagQueryExprType
 {
@@ -360,196 +264,4 @@ public enum GameplayTagQueryExprType
 }
 
 
-[GlobalClass]
-public partial class GameplayTagQueryExpression : Resource
-{
-    public GameplayTagQueryExprType ExprType { get; set; } = GameplayTagQueryExprType.Undefined;
-    public List<GameplayTagQueryExpression> ExprSet { get; set; } = new();
-    public List<GameplayTag> TagSet { get; set; } = new();
 
-    public bool IsValid()
-    {
-        if (ExprType == GameplayTagQueryExprType.Undefined)
-        {
-            return false;
-        }
-        else
-        {
-            if (ExprSet.Count == 0 && TagSet.Count == 0)
-            {
-                return false;
-            }
-
-            return true;
-        }
-    }
-
-    public GameplayTagQueryExpression AnyTagsMatch()
-    {
-        ExprType = GameplayTagQueryExprType.AnyTagsMatch;
-        return this;
-    }
-    
-    public GameplayTagQueryExpression AllTagsMatch()
-    {
-        ExprType = GameplayTagQueryExprType.AllTagsMatch;
-        return this;
-    }
-    
-    public GameplayTagQueryExpression NoTagsMatch()
-    {
-        ExprType = GameplayTagQueryExprType.NoTagsMatch;
-        return this;
-    }
-    
-    public GameplayTagQueryExpression AnyExprMatch()
-    {
-        ExprType = GameplayTagQueryExprType.AnyExprMatch;
-        return this;
-    }
-    
-    public GameplayTagQueryExpression AllExprMatch()
-    {
-        ExprType = GameplayTagQueryExprType.AllExprMatch;
-        return this;
-    }
-    
-    public GameplayTagQueryExpression NoExprMatch()
-    {
-        ExprType = GameplayTagQueryExprType.NoExprMatch;
-        return this;
-    }
-
-    //是否使用tag设置来判断
-    public bool UsesTagSet()
-    {
-        return ExprType == GameplayTagQueryExprType.AnyTagsMatch ||
-               ExprType == GameplayTagQueryExprType.AllTagsMatch ||
-               ExprType == GameplayTagQueryExprType.NoTagsMatch;
-    }
-
-    public GameplayTagQueryExpression AddTag(GameplayTag tag)
-    {
-        TagSet.Add(tag);
-        return this;
-    }
-
-    public GameplayTagQueryExpression AddTags(GameplayTagContainer tags)
-    {
-        for (var i = 0; i < tags.GameplayTags.Count; i++)
-        {
-            TagSet.Add(tags.GameplayTags[i]);
-        }
-        return this;
-    }
-
-    ///是否使用表达式expr来判断
-    public bool UsesExprSet()
-    {
-        return ExprType == GameplayTagQueryExprType.AnyExprMatch ||
-               ExprType == GameplayTagQueryExprType.AllExprMatch ||
-               ExprType == GameplayTagQueryExprType.NoExprMatch;
-    }
-
-    public GameplayTagQueryExpression AddExpr(GameplayTagQueryExpression expression)
-    {
-        ExprSet.Add(expression);
-        return this;
-    }
-
-    public bool Matches(GameplayTagContainer container)
-    {
-        switch (ExprType)
-        {
-            case GameplayTagQueryExprType.AnyTagsMatch:
-                for (var i = 0; i < TagSet.Count; i++)
-                {
-                    var tag = TagSet[i];
-                    if (container.HasTag(tag))
-                    {
-                        return true;
-                    }
-                }
-                return false;
-            case GameplayTagQueryExprType.AllTagsMatch:
-                for (var i = 0; i < TagSet.Count; i++)
-                {
-                    var tag = TagSet[i];
-                    if (!container.HasTag(tag))
-                    {
-                        return false;
-                    }
-                }
-                return true;
-            case GameplayTagQueryExprType.NoTagsMatch:
-                for (var i = 0; i < TagSet.Count; i++)
-                {
-                    var tag = TagSet[i];
-                    if (container.HasTag(tag))
-                    {
-                        return false;
-                    }
-                }
-                return true;
-            case GameplayTagQueryExprType.AnyExprMatch:
-                for (var i = 0; i < ExprSet.Count; i++)
-                {
-                    var expr = ExprSet[i];
-                    if (expr.Matches(container))
-                    {
-                        return true;
-                    }
-                }
-                return false;
-            case GameplayTagQueryExprType.AllExprMatch:
-                for (var i = 0; i < ExprSet.Count; i++)
-                {
-                    var expr = ExprSet[i];
-                    if (!expr.Matches(container))
-                    {
-                        return false;
-                    }
-                }
-                return true;
-            case GameplayTagQueryExprType.NoExprMatch:
-                for (var i = 0; i < ExprSet.Count; i++)
-                {
-                    var expr = ExprSet[i];
-                    if (expr.Matches(container))
-                    {
-                        return false;
-                    }
-                }
-                return true;
-            default:
-                return false;
-                
-        }
-    }
-}
-
-
-public class GameplayTagQuery
-{
-    public GameplayTagQueryExpression Expr { get; set; } = new();
-    public string Description { get; set; }
-
-    public bool IsEmpty()
-    {
-        return Expr.IsValid();
-    }
-
-    public static GameplayTagQuery Build(GameplayTagQueryExpression expression)
-    {
-        return new GameplayTagQuery()
-        {
-            Expr = expression,
-        };
-    }
-
-    //如果为空，我们认为就匹配任何标签
-    public bool Matches(GameplayTagContainer container)
-    {
-        return IsEmpty() || Expr.Matches(container);
-    }
-}
